Improve the verification of multiple permissions for projects and jobs from appconn

+++ exchangis-project/exchangis-project-server/src/main/java/com/webank/wedatasphere/exchangis/project/server/restful/external/ExchangisProjectDssAppConnRestfulApi.java
@@ -1,8 +1,10 @@
 package com.webank.wedatasphere.exchangis.project.server.restful.external;
 
 import com.webank.wedatasphere.exchangis.common.validator.groups.UpdateGroup;
+import com.webank.wedatasphere.exchangis.project.server.domain.OperationType;
 import com.webank.wedatasphere.exchangis.project.server.service.ProjectService;
 import com.webank.wedatasphere.exchangis.project.server.utils.ExchangisProjectRestfulUtils;
+import com.webank.wedatasphere.exchangis.project.server.utils.ProjectAuthorityUtils;
 import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectAppVo;
 import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectInfo;
 import org.apache.commons.math3.util.Pair;
@@ -75,19 +77,10 @@ public class ExchangisProjectDssAppConnRestfulApi {
         String username = SecurityFilter.getLoginUsername(request);
         try {
             ExchangisProjectInfo projectStored = projectService.getProjectById(Long.valueOf(projectVo.getId()));
-
-            if (!hasAuthority(username, projectStored)) {
+            if (!ProjectAuthorityUtils.hasProjectAuthority(username, projectStored, OperationType.PROJECT_ALTER)) {
                 return Message.error("You have no permission to update (没有项目的更新权限)");
             }
 
-//            if (!AuthorityUtils.hasOwnAuthority(Long.parseLong(projectVo.getId()), username) && !AuthorityUtils.hasEditAuthority(Long.parseLong(projectVo.getId()), username)) {
-//                return Message.error("You have no permission to update (没有编辑权限，无法更新项目)");
-//            }
-            String domain = projectStored.getDomain();
-            /*if (StringUtils.isNotBlank(domain) && !ExchangisProject.Domain.STANDALONE.name()
-                    .equalsIgnoreCase(domain)){
-                return Message.error("Cannot update the outer project (无法更新来自 " + domain + " 的外部项目)");
-            }*/
             LOG.info("UpdateProject vo: {}, userName: {}", JsonUtils.jackson().writeValueAsString(projectVo), username);
             projectService.updateProject(projectVo, username);
             return ExchangisProjectRestfulUtils.dealOk("更新工程成功",
@@ -110,14 +103,10 @@ public class ExchangisProjectDssAppConnRestfulApi {
         String username = SecurityFilter.getLoginUsername(request);
         try {
             ExchangisProjectInfo projectInfo = projectService.selectByName(name);
-//            if (!hasAuthority(username, projectInfo)){
-//                return Message.error("You have no permission to delete (删除工程失败)");
-//            }
-//            String domain = projectInfo.getDomain();
-//            if (StringUtils.isNotBlank(domain) && !ExchangisProject.Domain.STANDALONE.name()
-//                    .equalsIgnoreCase(domain)){
-//                return Message.error("Cannot delete the outer project (无法删除来自 " + domain + " 的外部项目)");
-//            }
+            if (!ProjectAuthorityUtils.hasProjectAuthority(username, projectInfo, OperationType.PROJECT_ALTER)) {
+                return Message.error("You have no permission to delete (删除项目失败)");
+            }
+
             projectService.deleteProjectByName(name);
             return ExchangisProjectRestfulUtils.dealOk("删除工程成功");
         } catch (Exception t) {
@@ -139,23 +128,15 @@ public class ExchangisProjectDssAppConnRestfulApi {
         String username = SecurityFilter.getLoginUsername(request);
         try {
             ExchangisProjectInfo projectInfo = projectService.selectByName(name);
+            if (!ProjectAuthorityUtils.hasProjectAuthority(username, projectInfo, OperationType.PROJECT_QUERY)){
+                return Message.error("You have no permission to query (没有项目查看权限)");
+            }
             return ExchangisProjectRestfulUtils.dealOk("根据名字获取工程成功",
                     new Pair<>("projectInfo",projectInfo));
         } catch (Exception t) {
             LOG.error("Failed to delete project for user {}", username, t);
             return Message.error("Failed to delete project (根据名字获取工程失败)");
         }
-
-
     }
 
-
-    /**
-     * @param username username
-     * @param project project
-     * @return
-     */
-    private boolean hasAuthority(String username, ExchangisProjectInfo project){
-        return Objects.nonNull(project) && username.equals(project.getCreateUser());
-    }
 }
